use vps' link interception opt-out
