+++ bifrost/package.json
@@ -83,14 +83,14 @@
     "react-dom": "^18.x.x",
     "tough-cookie": "^4.1.2",
     "uuid": "^9.0.0",
-    "vite-plugin-ssr": "0.4.130"
+    "vite-plugin-ssr": "0.4.131-commit-35ca471f7"
   },
   "peerDependencies": {
     "react": "^18.x.x",
     "react-dom": "^18.x.x",
     "typescript": ">=4.7",
     "jsdom": "^22.1.0",
-    "vite-plugin-ssr": "0.4.130"
+    "vite-plugin-ssr": "0.4.131-commit-35ca471f7"
   },
   "devDependencies": {
     "@types/node": "^18.11.9",
