+++ bifrost/proxy/pages/restorationVisit/onRenderClient.tsx
@@ -3,7 +3,8 @@ import { PageContextProxyRestorationVisit } from "../../../types/internal.js";
 import { PageShell } from "../../../lib/PageShell.js";
 import { renderReact } from "../../../lib/renderReact.js";
 import { Turbolinks } from "../../../lib/turbolinks/index.js";
-import { copyElementAttributes, getElementAttributes } from "../../../lib/turbolinks/util.js";
+import { getElementAttributes } from "../../../lib/getElementAttributes.js";
+import { copyElementAttributes } from "../../../lib/turbolinks/util.js";
 
 export default async function onRenderClient(
   pageContext: PageContextProxyRestorationVisit
