+++ tests/vite/package.json
@@ -20,7 +20,7 @@
     "ts-node": "^10.9.1",
     "uuid": "^9.0.0",
     "vite": "^4.3.2",
-    "vite-plugin-ssr": "0.4.130"
+    "vite-plugin-ssr": "0.4.131-commit-35ca471f7"
   },
   "devDependencies": {
     "@types/uuid": "^9.0.1"
